refactor(helm): split repo.go into auth.go + oci_chart.go
Same code-org symmetry as the git (#277) and bucket (#279) splits.
auth_test.go and oci_chart_test.go already existed without
production-side counterparts; the implementations lived in repo.go
alongside the HelmRepository index-fetch path.

- auth.go (new, 126 lines): helmRepoAuthOptions +
  helmRepoTLSOptions. Covered by auth_test.go's 10 tests.
- oci_chart.go (new, 233 lines): locateOCIChart +
  ociChartPathFromArtifact + chartSubdirStatus enum +
  findChartSubdir + chartYamlFilename / copiedOCILayerFilename
  constants + ociPullRef + fetchOCIChart + safeName. Covered by
  oci_chart_test.go's 7 tests.
- repo.go (518 → 195 lines): HelmRepository index fetch path —
  writeAtomic + ChartLoadResult + locateLocalChart +
  locateHelmRepoChart + fetchIndex + absChartURL.

Editing the OCI-chart-resolution surface no longer pulls the
HelmRepository index code into the editor scope. Behavior unchanged;
full suite + lint green.
